much cleaner and simpler accounts list page
