+++ src/app/modules/wallet/components/account-actions/account-actions.component.html
@@ -1,0 +1,15 @@
+<div class="flex items-center">
+  <button mat-raised-button color="accent" class="large-btn">
+    <span class="flex items-center">
+      <span class="mr-2">Import</span>
+      <mat-icon>cloud_upload</mat-icon>
+    </span>
+  </button>
+  <div class="mx-2"></div>
+  <button mat-raised-button color="primary" class="large-btn">
+    <span class="flex items-center">
+      <span class="mr-2">New Account</span>
+      <mat-icon>add</mat-icon>
+    </span>
+  </button>
+</div>
