+++ src/app/modules/wallet/components/account-actions/account-actions.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-account-actions',
+  templateUrl: './account-actions.component.html',
+  styles: [
+  ]
+})
+export class AccountActionsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
