+++ src/app/modules/wallet/components/account-selections/account-selections.component.ts
@@ -1,0 +1,14 @@
+import { SelectionModel } from '@angular/cdk/collections';
+import { Component, Input } from '@angular/core';
+import { TableData } from '../accounts-table/accounts-table.component';
+
+@Component({
+  selector: 'app-account-selections',
+  templateUrl: './account-selections.component.html',
+  styles: [
+  ]
+})
+export class AccountSelectionsComponent {
+  @Input() selection: SelectionModel<TableData> | null = null;
+  constructor() { }
+}
