--- src/app/modules/wallet/pages/account-list/account-list.component.html
@@ -1,88 +1,0 @@
-<div class="m-sm-30">
-  <app-breadcrumb></app-breadcrumb>
-  <div class="mb-8">
-    <div class="text-2xl mb-2 text-white leading-snug">
-      Your Validator Accounts List
-    </div>
-    <p class="m-0 text-muted text-base leading-snug">
-      Full list of all validating public keys managed by your Prysm wallet
-    </p>
-  </div>
-  <div class="relative flex justify-start flex-wrap items-center md:justify-between mb-4">
-    <mat-form-field class="search-bar text-base" appearance="fill">
-      <mat-label>Filter rows by public key, validator index, or account name</mat-label>
-      <input
-        matInput
-        placeholder="0x004a19ce..."
-        color="primary"
-        (keyup)="applySearchFilter($event)">
-      <mat-icon matSuffix>search</mat-icon>
-    </mat-form-field>
-    <div class="flex items-center">
-      <button mat-raised-button color="accent" class="large-btn">
-        <span class="flex items-center">
-          <span class="mr-2">Import</span>
-          <mat-icon>cloud_upload</mat-icon>
-        </span>
-      </button>
-      <div class="mx-2"></div>
-      <button mat-raised-button color="primary" class="large-btn">
-        <span class="flex items-center">
-          <span class="mr-2">New Account</span>
-          <mat-icon>add</mat-icon>
-        </span>
-      </button>
-    </div>
-  </div>
-
-  <div class="account-selections mb-6" *ngIf="selection?.selected.length">
-    <div class="text-muted text-lg">
-      Selected {{selection?.selected.length}} Accounts
-    </div>
-    <div class="my-6">
-      <mat-chip-list>
-        <mat-chip *ngFor="let account of selection?.selected">
-          {{account.publicKey | base64tohex | slice:0:16}}...
-        </mat-chip>
-      </mat-chip-list>
-    </div>
-    <div class="flex">
-      <button mat-stroked-button color="primary">
-        <span class="flex items-center">
-          <span class="mr-2">View in Block Explorer</span>
-          <mat-icon>open_in_new</mat-icon>
-        </span>
-      </button>
-      <div class="mx-2"></div>
-      <button mat-stroked-button color="accent">
-        <span class="flex items-center">
-          <span class="mr-2">Backup All</span>
-          <mat-icon>get_app</mat-icon>
-        </span>
-      </button>
-      <div class="mx-2"></div>
-      <button mat-stroked-button color="warn">
-        <span class="flex items-center">
-          <span class="mr-2">Delete All</span>
-          <mat-icon>delete_forever</mat-icon>
-        </span>
-      </button>
-    </div>
-  </div>
-
-  <div class="mat-elevation-z8 relative">
-    <div class="table-loading-shade" *ngIf="loading">
-      <mat-spinner></mat-spinner>
-    </div>
-
-    <div class="table-container bg-paper">
-      <app-accounts-table></app-accounts-table>
-    </div>
-
-    <mat-paginator 
-      (page)="handlePageEvent($event)" 
-      [length]="totalData"
-      [pageSizeOptions]="pageSizes">
-    </mat-paginator>
-  </div>
-</div>
