+++ src/app/modules/wallet/pages/accounts/accounts.component.html
@@ -1,0 +1,45 @@
+<div class="m-sm-30">
+  <app-breadcrumb></app-breadcrumb>
+  <div class="mb-8">
+    <div class="text-2xl mb-2 text-white leading-snug">
+      Your Validator Accounts List
+    </div>
+    <p class="m-0 text-muted text-base leading-snug">
+      Full list of all validating public keys managed by your Prysm wallet
+    </p>
+  </div>
+  <div
+    *ngIf="(tableDataSource$ | async) as dataSource"
+    class="relative flex justify-start flex-wrap items-center md:justify-between mb-4">
+    <mat-form-field class="search-bar text-base" appearance="fill">
+      <mat-label>Filter rows by public key, validator index, or account name</mat-label>
+      <input
+        matInput
+        placeholder="0x004a19ce..."
+        color="primary"
+        (keyup)="applySearchFilter($event, dataSource)">
+      <mat-icon matSuffix>search</mat-icon>
+    </mat-form-field>
+    <app-account-actions></app-account-actions>
+  </div>
+
+  <app-account-selections
+    [selection]="selection"></app-account-selections>
+
+  <div class="mat-elevation-z8 relative">
+    <div class="table-loading-shade" *ngIf="loading">
+      <mat-spinner></mat-spinner>
+    </div>
+    <div class="table-container bg-paper">
+      <app-accounts-table
+        *ngIf="(tableDataSource$ | async) as dataSource"
+        [dataSource]="dataSource"
+        [selection]="selection"></app-accounts-table>
+    </div>
+    <mat-paginator
+      (page)="handlePageEvent($event)"
+      [length]="totalData"
+      [pageSizeOptions]="pageSizes">
+    </mat-paginator>
+  </div>
+</div>
