+++ src/app/modules/wallet/wallet.module.ts
@@ -4,7 +4,7 @@ import { RouterModule } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { SharedModule } from '../shared/shared.module';
-import { AccountListComponent } from './pages/account-list/account-list.component';
+import { AccountsComponent } from './pages/accounts/accounts.component';
 import { WalletConfigComponent } from './components/wallet-config/wallet-config.component';
 import { IconTriggerSelectComponent } from './components/icon-trigger-select/icon-trigger-select.component';
 import { WalletDetailsComponent } from './pages/wallet-details/wallet-details.component';
@@ -12,10 +12,12 @@ import { WalletHelpComponent } from './components/wallet-help/wallet-help.compon
 import { WalletKindComponent } from './components/wallet-kind/wallet-kind.component';
 import { FilesAndDirectoriesComponent } from './components/files-and-directories/files-and-directories.component';
 import { AccountsTableComponent } from './components/accounts-table/accounts-table.component';
+import { AccountSelectionsComponent } from './components/account-selections/account-selections.component';
+import { AccountActionsComponent } from './components/account-actions/account-actions.component';
 
 @NgModule({
   declarations: [
-    AccountListComponent,
+    AccountsComponent,
     WalletConfigComponent,
     IconTriggerSelectComponent,
     WalletDetailsComponent,
@@ -23,6 +25,8 @@ import { AccountsTableComponent } from './components/accounts-table/accounts-tab
     WalletKindComponent,
     FilesAndDirectoriesComponent,
     AccountsTableComponent,
+    AccountSelectionsComponent,
+    AccountActionsComponent,
   ],
   imports: [
     CommonModule,
